Completely avoid moving windows that cannot be moved to a third of the display.
Previously Spectacle would try to move a window to a third of the display, regardless of whether the window can actually fit. If it couldn't Spectacle tries to move it back to where it came from. This behavior resulted in an annoying "flicker" as the window is moved around.

To prevent this issue Spectacle now attempts to remember which window sizes and positions have posed a problem in the past.

In next iterations Spectacle should persist this "blacklist" to the filesystem.